Fix the "-100.0 dB" issue
There was an issue caused where the display (and the host) would display
"-100.0 dB" instead of "-oo dB". It ultimately was caused by the display
rounding -99.96 to -100.0, but the value not ultimately reflecting that.
The "-oo" display is only triggered when the value is truly -100.0 dB.

Also, moved some of the gain parameter code into a different file, which
should later be moved to a different project.